add time interval for backtest period
