+++ action/0002_trade_decision.csv
@@ -1,141 +1,3 @@
 Date,Action
-2015-01-23,SELL
-2015-02-16,BUY
-2015-02-27,SELL
-2015-03-12,BUY
-2015-03-17,BUY
-2015-04-24,SELL
-2015-06-01,BUY
-2015-06-05,BUY
-2015-06-10,SELL
-2015-09-04,BUY
-2015-10-05,SELL
-2015-10-13,BUY
-2015-10-27,SELL
-2015-11-10,BUY
-2015-12-04,SELL
-2016-01-20,BUY
-2016-02-02,SELL
-2016-02-24,BUY
-2016-02-26,SELL
-2016-03-02,SELL
-2016-04-14,SELL
-2016-04-25,SELL
-2016-04-28,SELL
-2016-06-02,SELL
-2016-06-07,SELL
-2016-06-22,SELL
-2016-07-04,SELL
-2016-07-07,SELL
-2016-07-14,SELL
-2016-07-25,SELL
-2016-10-11,BUY
-2016-10-13,BUY
-2016-10-18,BUY
-2016-10-21,SELL
-2016-11-18,BUY
-2016-12-19,BUY
-2016-12-23,BUY
-2017-01-11,SELL
-2017-01-13,SELL
-2017-02-07,SELL
-2017-02-14,BUY
-2017-02-22,SELL
-2017-03-02,SELL
-2017-03-30,SELL
-2017-05-10,SELL
-2017-07-06,BUY
-2017-07-14,BUY
-2017-07-24,SELL
-2017-07-31,SELL
-2017-08-24,BUY
-2017-08-28,SELL
-2017-09-20,BUY
-2017-09-28,BUY
-2017-11-27,SELL
-2017-12-15,BUY
-2017-12-28,SELL
-2018-01-02,SELL
-2018-01-18,BUY
-2018-02-21,BUY
-2018-03-06,SELL
-2018-03-22,BUY
-2018-05-14,SELL
-2018-07-04,SELL
-2018-07-31,SELL
-2018-09-13,SELL
-2018-10-04,BUY
-2018-10-25,SELL
-2019-01-16,BUY
-2019-01-22,SELL
-2019-02-19,SELL
-2019-03-14,BUY
-2019-04-16,SELL
-2019-04-23,BUY
-2019-05-27,SELL
-2019-07-31,BUY
-2019-08-14,BUY
-2019-08-27,BUY
-2019-10-30,SELL
-2019-11-04,SELL
-2019-11-22,BUY
-2019-11-27,BUY
-2019-12-24,SELL
-2019-12-31,BUY
-2020-03-16,SELL
-2020-03-18,BUY
-2020-03-30,BUY
-2020-04-14,SELL
-2020-05-05,BUY
-2020-05-28,SELL
-2020-07-15,BUY
-2020-08-07,SELL
-2020-08-11,SELL
-2020-09-18,BUY
-2020-09-24,BUY
-2020-09-29,BUY
-2020-12-01,BUY
-2020-12-03,BUY
-2020-12-14,BUY
-2021-02-17,BUY
-2021-02-23,SELL
-2021-04-20,SELL
-2021-04-26,SELL
-2021-04-29,SELL
-2021-05-06,SELL
-2021-07-27,SELL
-2021-08-16,BUY
-2021-08-31,SELL
-2021-09-09,BUY
-2021-09-20,BUY
-2021-11-01,SELL
-2021-12-23,SELL
-2021-12-24,BUY
-2022-02-08,SELL
-2022-02-11,SELL
-2022-02-15,SELL
-2022-02-17,SELL
-2022-03-01,BUY
-2022-03-09,BUY
-2022-03-16,BUY
-2022-03-23,SELL
-2022-05-23,SELL
-2022-06-10,BUY
-2022-08-15,SELL
-2022-08-17,SELL
-2022-09-06,BUY
-2022-09-21,BUY
-2022-09-26,BUY
-2022-10-31,BUY
-2023-01-10,SELL
-2023-02-02,BUY
-2023-02-17,BUY
-2023-02-23,BUY
-2023-02-28,BUY
-2023-04-04,SELL
-2023-04-13,BUY
-2023-04-18,SELL
-2023-04-24,SELL
-2023-05-08,SELL
-2023-05-15,SELL
-2023-05-29,BUY
+2022-11-14,BUY
+2023-01-23,SELL
