+++ action/0005_trade_decision.csv
@@ -1,155 +1,4 @@
 Date,Action
-2015-01-21,BUY
-2015-03-06,BUY
-2015-04-08,SELL
-2015-05-08,SELL
-2015-06-01,BUY
-2015-06-29,BUY
-2015-07-02,BUY
-2015-07-10,SELL
-2015-07-14,BUY
-2015-08-20,BUY
-2015-09-30,BUY
-2015-10-20,SELL
-2015-11-02,BUY
-2015-12-10,BUY
-2016-01-08,BUY
-2016-02-24,BUY
-2016-04-05,BUY
-2016-04-21,SELL
-2016-04-25,SELL
-2016-05-12,BUY
-2016-05-18,BUY
-2016-06-15,BUY
-2016-07-18,SELL
-2016-07-25,SELL
-2016-08-01,BUY
-2016-08-31,SELL
-2016-12-07,SELL
-2016-12-29,BUY
-2017-01-18,SELL
-2017-01-26,SELL
-2017-02-07,SELL
-2017-02-13,SELL
-2017-02-20,SELL
-2017-02-28,BUY
-2017-03-07,BUY
-2017-03-27,SELL
-2017-05-09,SELL
-2017-05-11,BUY
-2017-05-22,SELL
-2017-06-08,SELL
-2017-06-29,SELL
-2017-07-25,SELL
-2017-07-31,SELL
-2017-08-08,SELL
-2017-08-18,BUY
-2017-09-27,SELL
-2017-10-03,SELL
-2017-10-09,SELL
-2017-10-30,BUY
-2017-11-20,SELL
-2017-12-04,BUY
-2017-12-13,SELL
-2017-12-28,SELL
-2018-01-09,SELL
-2018-02-06,BUY
-2018-03-08,BUY
-2018-03-13,BUY
-2018-04-03,BUY
-2018-04-24,SELL
-2018-06-25,BUY
-2018-07-24,SELL
-2018-08-13,BUY
-2018-10-15,BUY
-2018-11-15,SELL
-2018-12-03,SELL
-2019-02-01,SELL
-2019-02-13,BUY
-2019-02-18,SELL
-2019-03-05,BUY
-2019-03-08,BUY
-2019-03-13,SELL
-2019-04-12,BUY
-2019-04-15,SELL
-2019-05-03,SELL
-2019-05-24,BUY
-2019-06-04,SELL
-2019-07-31,BUY
-2019-09-11,SELL
-2019-10-24,SELL
-2019-11-11,BUY
-2019-11-13,BUY
-2019-12-19,SELL
-2019-12-23,SELL
-2019-12-24,BUY
-2020-01-21,BUY
-2020-01-29,BUY
-2020-01-31,SELL
-2020-02-27,BUY
-2020-04-16,BUY
-2020-05-29,BUY
-2020-06-26,BUY
-2020-07-24,BUY
-2020-08-03,BUY
-2020-09-01,BUY
-2020-09-14,BUY
-2020-10-23,SELL
-2020-11-04,SELL
-2020-11-16,BUY
-2020-12-02,SELL
-2020-12-23,BUY
-2021-01-12,SELL
-2021-01-19,SELL
-2021-02-17,SELL
-2021-03-29,BUY
-2021-04-30,SELL
-2021-05-06,SELL
-2021-06-18,BUY
-2021-06-28,BUY
-2021-07-06,SELL
-2021-07-07,BUY
-2021-07-19,BUY
-2021-08-13,SELL
-2021-08-30,BUY
-2021-09-14,BUY
-2021-09-20,BUY
-2021-10-07,SELL
-2021-10-26,BUY
-2021-11-15,BUY
-2021-11-30,BUY
-2021-12-17,SELL
-2022-01-04,SELL
-2022-01-26,SELL
-2022-02-04,SELL
-2022-02-14,SELL
-2022-02-28,BUY
-2022-03-04,BUY
-2022-03-15,BUY
-2022-03-25,SELL
-2022-04-26,BUY
-2022-05-06,SELL
-2022-05-12,BUY
-2022-05-30,SELL
-2022-06-01,SELL
-2022-06-20,BUY
-2022-07-04,SELL
-2022-07-15,BUY
-2022-07-19,BUY
-2022-08-10,SELL
-2022-08-17,BUY
-2022-08-24,BUY
-2022-08-26,BUY
-2022-09-02,SELL
-2022-09-26,BUY
-2022-10-07,BUY
-2022-11-14,SELL
-2022-12-12,BUY
-2023-01-04,SELL
-2023-03-14,BUY
-2023-03-16,BUY
-2023-04-12,SELL
-2023-04-18,SELL
-2023-05-02,SELL
-2023-05-10,SELL
-2023-05-12,SELL
+2022-11-14,BUY
+2023-01-02,SELL
+2023-02-13,BUY
