+++ action/0016_trade_decision.csv
@@ -1,168 +1,3 @@
 Date,Action
-2015-01-26,SELL
-2015-02-18,BUY
-2015-03-04,BUY
-2015-03-12,BUY
-2015-04-01,SELL
-2015-05-04,SELL
-2015-06-01,BUY
-2015-06-15,BUY
-2015-07-08,BUY
-2015-07-10,SELL
-2015-07-31,BUY
-2015-08-05,BUY
-2015-09-22,SELL
-2015-10-23,SELL
-2015-11-06,BUY
-2015-11-13,BUY
-2015-12-10,BUY
-2015-12-21,BUY
-2016-01-08,BUY
-2016-02-26,SELL
-2016-03-02,SELL
-2016-03-14,SELL
-2016-03-22,BUY
-2016-04-22,SELL
-2016-05-05,BUY
-2016-05-27,SELL
-2016-06-02,SELL
-2016-06-07,SELL
-2016-06-28,BUY
-2016-07-04,SELL
-2016-07-12,BUY
-2016-09-05,SELL
-2016-09-19,SELL
-2016-09-21,SELL
-2016-10-12,BUY
-2016-11-01,SELL
-2016-11-07,BUY
-2016-11-09,BUY
-2016-11-14,BUY
-2016-11-17,SELL
-2016-12-09,SELL
-2016-12-29,BUY
-2017-01-09,SELL
-2017-02-10,SELL
-2017-02-16,BUY
-2017-02-21,SELL
-2017-04-03,SELL
-2017-04-10,SELL
-2017-04-13,SELL
-2017-05-23,BUY
-2017-06-06,SELL
-2017-06-08,SELL
-2017-06-13,SELL
-2017-06-26,BUY
-2017-07-18,SELL
-2017-07-27,SELL
-2017-08-04,SELL
-2017-08-09,SELL
-2017-08-30,SELL
-2017-09-28,BUY
-2017-10-17,SELL
-2017-11-03,BUY
-2017-12-07,SELL
-2017-12-13,BUY
-2017-12-28,SELL
-2018-01-10,BUY
-2018-01-12,SELL
-2018-02-09,BUY
-2018-02-26,SELL
-2018-03-01,SELL
-2018-03-19,BUY
-2018-03-28,SELL
-2018-03-29,BUY
-2018-05-14,SELL
-2018-06-25,BUY
-2018-08-17,BUY
-2018-09-05,BUY
-2018-09-26,SELL
-2018-10-08,BUY
-2018-10-30,BUY
-2018-11-30,SELL
-2018-12-19,BUY
-2019-01-15,SELL
-2019-01-18,SELL
-2019-02-20,SELL
-2019-03-28,SELL
-2019-04-24,BUY
-2019-05-31,BUY
-2019-06-06,SELL
-2019-06-12,BUY
-2019-06-21,SELL
-2019-07-03,SELL
-2019-07-12,BUY
-2019-07-25,BUY
-2019-07-31,BUY
-2019-10-04,BUY
-2019-10-28,SELL
-2019-12-19,BUY
-2019-12-24,SELL
-2020-01-03,SELL
-2020-01-15,SELL
-2020-01-31,BUY
-2020-02-10,BUY
-2020-02-14,SELL
-2020-03-18,BUY
-2020-04-08,SELL
-2020-05-20,BUY
-2020-06-12,SELL
-2020-06-18,SELL
-2020-07-22,BUY
-2020-08-12,SELL
-2020-08-21,SELL
-2020-08-26,SELL
-2020-09-01,BUY
-2020-10-22,SELL
-2020-11-06,BUY
-2020-12-28,BUY
-2021-01-12,SELL
-2021-02-08,BUY
-2021-02-19,SELL
-2021-03-03,SELL
-2021-03-17,BUY
-2021-03-24,BUY
-2021-05-04,BUY
-2021-06-29,BUY
-2021-07-09,SELL
-2021-08-04,BUY
-2021-08-10,BUY
-2021-09-07,BUY
-2021-09-09,BUY
-2021-09-16,BUY
-2021-10-04,BUY
-2021-10-19,SELL
-2021-11-30,BUY
-2021-12-13,BUY
-2022-01-10,SELL
-2022-01-21,SELL
-2022-02-10,SELL
-2022-02-17,SELL
-2022-02-28,BUY
-2022-03-15,BUY
-2022-04-04,SELL
-2022-04-22,BUY
-2022-04-25,SELL
-2022-05-30,SELL
-2022-06-01,BUY
-2022-06-17,BUY
-2022-06-22,BUY
-2022-06-24,BUY
-2022-06-28,SELL
-2022-08-19,BUY
-2022-08-22,SELL
-2022-08-31,BUY
-2022-09-06,BUY
-2022-09-28,SELL
-2022-09-29,BUY
-2022-10-11,BUY
-2022-11-25,SELL
-2022-12-07,BUY
-2022-12-30,SELL
-2023-01-04,SELL
-2023-03-10,BUY
-2023-03-30,SELL
-2023-04-06,SELL
-2023-04-12,SELL
-2023-05-11,SELL
-2023-05-25,BUY
+2023-01-02,SELL
+2023-02-27,BUY
