+++ action/0700_trade_decision.csv
@@ -1,157 +1,2 @@
 Date,Action
-2015-01-21,SELL
-2015-02-05,SELL
-2015-03-30,SELL
-2015-06-01,BUY
-2015-06-05,BUY
-2015-07-08,BUY
-2015-07-24,SELL
-2015-08-06,BUY
-2015-08-24,BUY
-2015-09-02,BUY
-2015-10-14,SELL
-2015-10-29,BUY
-2015-11-03,SELL
-2015-12-11,BUY
-2016-01-07,BUY
-2016-01-20,BUY
-2016-01-21,SELL
-2016-03-02,SELL
-2016-03-14,SELL
-2016-03-21,BUY
-2016-04-07,SELL
-2016-04-19,SELL
-2016-05-04,BUY
-2016-05-12,BUY
-2016-05-25,SELL
-2016-05-27,SELL
-2016-06-07,SELL
-2016-07-15,SELL
-2016-07-25,SELL
-2016-08-18,SELL
-2016-08-30,SELL
-2016-09-06,SELL
-2016-09-23,SELL
-2016-10-31,BUY
-2016-11-09,BUY
-2016-11-14,BUY
-2016-11-24,SELL
-2016-12-12,BUY
-2016-12-15,BUY
-2017-01-09,SELL
-2017-01-26,SELL
-2017-02-07,SELL
-2017-02-08,BUY
-2017-02-16,SELL
-2017-03-17,SELL
-2017-04-24,SELL
-2017-07-24,SELL
-2017-09-18,SELL
-2017-09-27,SELL
-2017-10-18,SELL
-2017-11-06,SELL
-2017-12-28,SELL
-2018-01-04,SELL
-2018-01-12,SELL
-2018-01-17,SELL
-2018-02-12,BUY
-2018-04-03,BUY
-2018-04-12,BUY
-2018-04-16,SELL
-2018-06-06,SELL
-2018-06-26,BUY
-2018-07-03,BUY
-2018-07-27,BUY
-2018-08-15,BUY
-2018-09-11,BUY
-2018-09-17,BUY
-2018-10-08,BUY
-2018-10-15,BUY
-2018-10-26,BUY
-2018-11-30,SELL
-2018-12-31,BUY
-2019-01-04,SELL
-2019-01-07,BUY
-2019-01-10,SELL
-2019-01-22,SELL
-2019-02-01,SELL
-2019-02-11,SELL
-2019-03-20,SELL
-2019-04-11,SELL
-2019-05-20,BUY
-2019-05-31,SELL
-2019-07-04,SELL
-2019-07-26,BUY
-2019-07-29,SELL
-2019-08-07,SELL
-2019-08-13,BUY
-2019-09-13,SELL
-2019-09-17,SELL
-2019-10-02,BUY
-2019-12-13,BUY
-2020-01-13,SELL
-2020-02-14,SELL
-2020-02-19,SELL
-2020-03-16,BUY
-2020-03-31,SELL
-2020-04-07,SELL
-2020-04-14,SELL
-2020-04-17,BUY
-2020-04-22,SELL
-2020-04-27,SELL
-2020-06-17,SELL
-2020-10-14,SELL
-2020-10-19,SELL
-2020-11-05,SELL
-2020-12-28,BUY
-2021-01-18,SELL
-2021-02-01,SELL
-2021-02-19,SELL
-2021-02-26,BUY
-2021-03-09,BUY
-2021-03-23,BUY
-2021-03-31,BUY
-2021-04-07,SELL
-2021-05-13,BUY
-2021-06-22,BUY
-2021-07-08,BUY
-2021-07-26,BUY
-2021-08-03,BUY
-2021-09-06,SELL
-2021-09-29,BUY
-2021-11-09,BUY
-2021-12-06,BUY
-2021-12-09,BUY
-2021-12-13,SELL
-2022-02-10,SELL
-2022-02-17,BUY
-2022-02-24,BUY
-2022-04-04,SELL
-2022-04-27,BUY
-2022-06-09,SELL
-2022-06-14,SELL
-2022-06-30,BUY
-2022-07-12,BUY
-2022-07-27,BUY
-2022-08-09,BUY
-2022-08-26,SELL
-2022-09-02,SELL
-2022-09-15,BUY
-2022-10-11,BUY
-2022-10-17,BUY
-2022-10-20,BUY
-2022-10-24,BUY
-2022-11-11,SELL
-2022-11-23,SELL
-2022-11-29,SELL
-2022-12-08,BUY
-2022-12-09,SELL
-2022-12-14,SELL
-2023-01-09,SELL
-2023-02-28,BUY
-2023-03-10,BUY
-2023-03-30,SELL
-2023-04-06,SELL
-2023-04-21,BUY
-2023-05-04,BUY
-2023-05-08,BUY
+2023-01-16,SELL
