+++ back_test_data/0002_back_test.csv
@@ -1,74 +1,3 @@
 Date,Cash,Number of Stock,Stock Price,Total value
-2015-02-16,1.60113525390625,372.0,67.19999694824219,25000.0
-2015-02-27,25948.601135253906,0,69.75,25948.601135253906
-2015-03-12,64.04991149902344,401.0,64.55000305175781,25948.601135253906
-2015-04-24,27532.549911499023,0,68.5,27532.549911499023
-2015-06-01,19.35053253173828,407.0,67.5999984741211,27532.549911499023
-2015-06-10,26413.299911499023,0,64.8499984741211,26413.299911499023
-2015-09-04,18.099266052246094,423.0,62.400001525878906,26413.299911499023
-2015-10-05,28359.099266052246,0,67.0,28359.099266052246
-2015-10-13,51.09862518310547,420.0,67.4000015258789,28359.099266052246
-2015-10-27,28989.099266052246,0,68.9000015258789,28989.099266052246
-2015-11-10,40.300621032714844,444.0,65.19999694824219,28989.099266052246
-2015-12-04,28878.099266052246,0,64.94999694824219,28878.099266052246
-2016-01-20,3.0992660522460938,462.0,62.5,28878.099266052246
-2016-02-02,30495.099266052246,0,66.0,30495.099266052246
-2016-02-24,55.599266052246094,446.0,68.25,30495.099266052246
-2016-02-26,30584.29790496826,0,68.44999694824219,30584.29790496826
-2016-10-11,31.048507690429688,395.0,77.3499984741211,30584.29790496826
-2016-10-21,30920.047302246094,0,78.19999694824219,30920.047302246094
-2016-11-18,39.74858093261719,419.0,73.69999694824219,30920.047302246094
-2017-01-11,31674.248580932617,0,75.5,31674.248580932617
-2017-02-14,43.74919891357422,405.0,78.0999984741211,31674.248580932617
-2017-02-22,31552.75043487549,0,77.80000305175781,31552.75043487549
-2017-07-06,66.54984283447266,388.0,81.1500015258789,31552.75043487549
-2017-07-24,32406.349250793457,0,83.3499984741211,32406.349250793457
-2017-08-24,29.04985809326172,398.0,81.3499984741211,32406.349250793457
-2017-08-28,32605.349250793457,0,81.8499984741211,32605.349250793457
-2017-09-20,63.59986877441406,405.0,80.3499984741211,32605.349250793457
-2017-11-27,33071.099868774414,0,81.5,33071.099868774414
-2017-12-15,49.09986877441406,418.0,79.0,33071.099868774414
-2017-12-28,33530.89923095703,0,80.0999984741211,33530.89923095703
-2018-01-18,47.29988098144531,426.0,78.5999984741211,33530.89923095703
-2018-03-06,33658.70053100586,0,78.9000015258789,33658.70053100586
-2018-03-22,43.700531005859375,415.0,81.0,33658.70053100586
-2018-05-14,34509.45179748535,0,83.05000305175781,34509.45179748535
-2018-10-04,82.95179748535156,389.0,88.5,34509.45179748535
-2018-10-25,33848.15298461914,0,86.80000305175781,33848.15298461914
-2019-01-16,67.15185546875728,370.0,91.3000030517578,33848.15298461914
-2019-01-22,33700.15242004395,0,90.9000015258789,33700.15242004395
-2019-03-14,84.252983093269,369.0,91.0999984741211,33700.15242004395
-2019-04-16,32999.051856994636,0,89.19999694824219,32999.051856994636
-2019-04-23,7.202995300300245,373.0,88.44999694824219,32999.051856994636
-2019-05-27,34584.30185699464,0,92.6999969482422,34584.30185699464
-2019-07-31,58.05185699464346,405.0,85.25,34584.30185699464
-2019-10-30,33166.80185699464,0,81.75,33166.80185699464
-2019-11-22,44.20311737062002,413.0,80.19999694824219,33166.80185699464
-2019-12-24,34054.75248718263,0,82.3499984741211,34054.75248718263
-2019-12-31,66.25185394288565,415.0,81.9000015258789,34054.75248718263
-2020-03-16,32270.25122070314,0,77.5999984741211,32270.25122070314
-2020-03-18,43.05187988282705,432.0,74.5999984741211,32270.25122070314
-2020-04-14,35164.6531982422,0,81.30000305175781,35164.6531982422
-2020-05-05,60.15452575685049,435.0,80.69999694824219,35164.6531982422
-2020-05-28,33228.90452575685,0,76.25,33228.90452575685
-2020-07-15,36.65591430665518,455.0,72.94999694824219,33228.90452575685
-2020-08-07,33888.65660858156,0,74.4000015258789,33888.65660858156
-2020-09-18,32.657310485854396,460.0,73.5999984741211,33888.65660858156
-2021-02-23,34624.65590667726,0,75.19999694824219,34624.65590667726
-2021-08-16,28.207252502455958,441.0,78.44999694824219,34624.65590667726
-2021-08-31,34315.957252502456,0,77.75,34315.957252502456
-2021-09-09,3.457252502455958,450.0,76.25,34315.957252502456
-2021-11-01,34293.45587921144,0,76.19999694824219,34293.45587921144
-2021-12-24,41.85721588136221,438.0,78.19999694824219,34293.45587921144
-2022-02-08,35016.1565475464,0,79.8499984741211,35016.1565475464
-2022-03-01,73.65586090089346,450.0,77.6500015258789,35016.1565475464
-2022-03-23,34071.15723419191,0,75.55000305175781,34071.15723419191
-2022-06-10,59.40723419190908,449.0,75.75,34071.15723419191
-2022-08-15,30434.257919311538,0,67.6500015258789,30434.257919311538
-2022-09-06,46.508628845229396,465.0,65.3499984741211,30434.257919311538
-2023-01-10,26830.507919311534,0,57.59999847412109,26830.507919311534
-2023-02-02,47.85861968995596,459.0,58.34999847412109,26830.507919311534
-2023-04-04,26233.808269500747,0,57.04999923706055,26233.808269500747
-2023-04-13,50.40861511231924,453.0,57.79999923706055,26233.808269500747
-2023-04-18,26369.70792388917,0,58.09999847412109,26369.70792388917
-2023-05-29,46.05862426759268,459.0,57.34999847412109,26369.70792388917
+2022-11-14,3.5996665954589844,437.0,57.20000076293945,25000.0
+2023-01-23,25284.048999786373,0,57.84999847412109,25284.048999786373
