+++ back_test_data/0005_back_test.csv
@@ -1,81 +1,4 @@
 Date,Cash,Number of Stock,Stock Price,Total value
-2015-01-21,7.6005401611328125,354.0,70.5999984741211,25000.0
-2015-04-08,24168.100540161133,0,68.25,24168.100540161133
-2015-06-01,19.151039123535156,327.0,73.8499984741211,24168.100540161133
-2015-07-10,22287.850540161133,0,68.0999984741211,22287.850540161133
-2015-07-14,9.09954833984375,325.0,68.55000305175781,22287.850540161133
-2015-10-20,20256.59930038452,0,62.29999923706055,20256.59930038452
-2015-11-02,56.09955596923828,335.0,60.29999923706055,20256.59930038452
-2016-02-24,7.399555206298828,336.0,48.70000076293945,16370.599811553955
-2016-04-21,17781.80006790161,0,52.900001525878906,17781.80006790161
-2016-05-12,0.20035171508789062,372.0,47.79999923706055,17781.80006790161
-2016-07-18,18507.200351715088,0,49.75,18507.200351715088
-2016-08-01,16.000629425048828,364.0,50.79999923706055,18507.200351715088
-2016-08-31,20946.00062942505,0,57.5,20946.00062942505
-2016-12-29,6.900371551513672,338.0,61.95000076293945,20946.00062942505
-2017-01-18,21875.499340057373,0,64.69999694824219,21875.499340057373
-2017-02-28,0.4993400573730469,350.0,62.5,21875.499340057373
-2017-03-27,21962.999340057373,0,62.75,21962.999340057373
-2017-05-11,15.150325775146484,323.0,67.94999694824219,21962.999340057373
-2017-05-22,21882.249340057373,0,67.69999694824219,21882.249340057373
-2017-08-18,37.44843673706055,296.0,73.80000305175781,21882.249340057373
-2017-09-27,22592.647533416748,0,76.19999694824219,22592.647533416748
-2017-10-30,75.59798049926758,293.0,76.8499984741211,22592.647533416748
-2017-11-20,22079.897533416748,0,75.0999984741211,22079.897533416748
-2017-12-04,13.098400115966797,284.0,77.69999694824219,22079.897533416748
-2017-12-13,22718.897533416748,0,79.94999694824219,22718.897533416748
-2018-02-06,50.597965240478516,283.0,80.0999984741211,22718.897533416748
-2018-04-24,22025.54839706421,0,77.6500015258789,22025.54839706421
-2018-06-25,48.048397064208984,298.0,73.75,22025.54839706421
-2018-07-24,22293.748851776123,0,74.6500015258789,22293.748851776123
-2018-08-13,14.849323272705078,309.0,72.0999984741211,22293.748851776123
-2018-11-15,20408.849323272705,0,66.0,20408.849323272705
-2019-02-13,10.449787139892578,304.0,67.0999984741211,20408.849323272705
-2019-02-18,20591.248859405518,0,67.69999694824219,20591.248859405518
-2019-03-05,63.59934616088867,319.0,64.3499984741211,20591.248859405518
-2019-03-13,20607.19983291626,0,64.4000015258789,20607.19983291626
-2019-04-12,1.9993629455566406,308.0,66.9000015258789,20607.19983291626
-2019-04-15,20668.798892974854,0,67.0999984741211,20668.798892974854
-2019-05-24,2.3984107971191406,316.0,65.4000015258789,20668.798892974854
-2019-06-04,20131.598651885986,0,63.70000076293945,20131.598651885986
-2019-07-31,49.64913558960325,317.0,63.34999847412109,20131.598651885986
-2019-09-11,19307.399135589603,0,60.75,19307.399135589603
-2019-11-11,14.399135589603247,327.0,59.0,19307.399135589603
-2019-12-19,19895.998886108402,0,60.79999923706055,19895.998886108402
-2019-12-24,51.998886108402075,328.0,60.5,19895.998886108402
-2020-01-31,18698.798385620117,0,56.84999847412109,18698.798385620117
-2020-02-27,39.597862243652344,343.0,54.400001525878906,18698.798385620117
-2020-05-29,3.74786376953125,344.0,35.849998474121094,12336.147338867188
-2020-10-23,11097.747863769531,0,32.25,11097.747863769531
-2020-11-16,0.9480819702148438,286.0,38.79999923706055,11097.747863769531
-2020-12-02,12155.948081970215,0,42.5,12155.948081970215
-2020-12-23,14.098316192626953,307.0,39.54999923706055,12155.948081970215
-2021-01-12,13291.848316192627,0,43.25,13291.848316192627
-2021-03-29,32.44876480102721,294.0,45.09999847412109,13291.848316192627
-2021-04-30,14394.348316192627,0,48.84999847412109,14394.348316192627
-2021-06-18,45.5480842590332,304.0,47.20000076293945,14394.348316192627
-2021-07-06,13801.548084259033,0,45.25,13801.548084259033
-2021-07-07,20.148555755617053,309.0,44.59999847412109,13801.548084259033
-2021-08-13,13677.948791503908,0,44.20000076293945,13677.948791503908
-2021-08-30,16.74829101562682,328.0,41.650001525878906,13677.948791503908
-2021-10-07,14202.748291015627,0,43.25,14202.748291015627
-2021-10-26,0.248291015626819,299.0,47.5,14202.748291015627
-2021-12-17,13873.84874725342,0,46.400001525878906,13873.84874725342
-2022-02-28,24.249137878421607,256.0,54.09999847412109,13873.84874725342
-2022-03-25,13541.048942565922,0,52.79999923706055,13541.048942565922
-2022-04-26,27.548942565921607,273.0,49.5,13541.048942565922
-2022-05-06,13404.548942565922,0,49.0,13404.548942565922
-2022-05-12,34.04894256592161,286.0,46.75,13404.548942565922
-2022-05-30,15020.449378967289,0,52.400001525878906,15020.449378967289
-2022-06-20,6.799610137943091,303.0,49.54999923706055,15020.449378967289
-2022-07-04,15535.549610137943,0,51.25,15535.549610137943
-2022-07-15,33.049362182620825,325.0,47.70000076293945,15535.549610137943
-2022-08-10,16916.799610137943,0,51.95000076293945,16916.799610137943
-2022-08-17,33.049362182620825,325.0,51.95000076293945,16916.799610137943
-2022-09-02,15665.548866271974,0,48.09999847412109,15665.548866271974
-2022-09-26,33.048866271974475,370.0,42.25,15665.548866271974
-2022-11-14,16183.54943084717,0,43.650001525878906,16183.54943084717
-2022-12-12,47.149948120120825,339.0,47.59999847412109,16183.54943084717
-2023-01-04,16997.14994812012,0,50.0,16997.14994812012
-2023-03-14,43.74946594238645,316.0,53.650001525878906,16997.14994812012
-2023-04-12,17344.749465942386,0,54.75,17344.749465942386
+2022-11-14,11.099136352539062,566.0,44.150001525878906,25000.0
+2023-01-02,29556.29956817627,0,52.20000076293945,29556.29956817627
+2023-02-13,48.8991813659668,507.0,58.20000076293945,29556.29956817627
