+++ back_test_data/0016_back_test.csv
@@ -1,96 +1,2 @@
 Date,Cash,Number of Stock,Stock Price,Total value
-2015-02-18,73.19969177246094,202.0,123.4000015258789,25000.0
-2015-04-01,24212.19969177246,0,119.5,24212.19969177246
-2015-06-01,24.400802612304688,182.0,132.89999389648438,24212.19969177246
-2015-07-10,22010.001358032223,0,120.8000030517578,22010.001358032223
-2015-07-31,95.60163879394167,184.0,119.0999984741211,22010.001358032223
-2015-09-22,19397.20191955566,0,104.9000015258789,19397.20191955566
-2015-11-06,39.30133056640625,193.0,100.3000030517578,19397.20191955566
-2016-02-26,17023.301330566406,0,88.0,17023.301330566406
-2016-03-22,80.95105743408203,179.0,94.6500015258789,17023.301330566406
-2016-04-22,17730.350784301758,0,98.5999984741211,17730.350784301758
-2016-05-05,22.35136413573855,190.0,93.1999969482422,17730.350784301758
-2016-05-27,16979.85136413574,0,89.25,16979.85136413574
-2016-06-28,74.25077819823855,192.0,88.05000305175781,16979.85136413574
-2016-07-04,18650.25077819824,0,96.75,18650.25077819824
-2016-07-12,68.05105590819949,182.0,102.0999984741211,18650.25077819824
-2016-09-05,21380.25077819824,0,117.0999984741211,21380.25077819824
-2016-10-12,98.65135192870366,188.0,113.1999969482422,21380.25077819824
-2016-11-01,22357.851638793938,0,118.4000015258789,22357.851638793938
-2016-11-07,101.85163879393804,214.0,104.0,22357.851638793938
-2016-11-14,2.1016387939380365,215.0,99.75,21448.351638793938
-2016-11-17,21717.101638793938,0,101.0,21717.101638793938
-2016-12-29,94.30130004882085,222.0,97.4000015258789,21717.101638793938
-2017-01-09,23182.30130004882,0,104.0,23182.30130004882
-2017-02-16,6.100646972652612,214.0,108.3000030517578,23182.30130004882
-2017-02-21,23909.899993896484,0,111.6999969482422,23909.899993896484
-2017-05-23,45.800315856933594,211.0,113.0999984741211,23909.899993896484
-2017-06-06,26230.899993896484,0,124.0999984741211,26230.899993896484
-2017-06-26,18.399993896484375,225.0,116.5,26230.899993896484
-2017-07-18,26838.39930725098,0,119.1999969482422,26838.39930725098
-2017-09-28,62.79866027832759,212.0,126.3000030517578,26838.39930725098
-2017-10-17,28301.19801330567,0,133.1999969482422,28301.19801330567
-2017-11-03,13.198013305671338,221.0,128.0,28301.19801330567
-2017-12-07,26930.998687744144,0,121.8000030517578,26930.998687744144
-2017-12-13,7.998687744144263,218.0,123.5,26930.998687744144
-2017-12-28,28326.197357177738,0,129.89999389648438,28326.197357177738
-2018-01-10,108.99606323242551,212.0,133.10000610351562,28326.197357177738
-2018-01-12,28304.996063232426,0,133.0,28304.996063232426
-2018-02-09,83.99676513671875,230.0,122.6999969482422,28304.996063232426
-2018-02-26,30121.998168945312,0,130.60000610351562,30121.998168945312
-2018-03-19,124.3974609375,232.0,129.3000030517578,30121.998168945312
-2018-03-28,29077.99816894531,0,124.8000030517578,29077.99816894531
-2018-03-29,61.99816894530886,234.0,124.0,29077.99816894531
-2018-05-14,30341.596740722653,0,129.39999389648438,30341.596740722653
-2018-06-25,0.796356201168237,252.0,120.4000015258789,30341.596740722653
-2018-09-26,29207.596740722653,0,115.9000015258789,29207.596740722653
-2018-10-08,22.396331787105737,268.0,108.9000015258789,29207.596740722653
-2018-11-30,29904.396331787106,0,111.5,29904.396331787106
-2018-12-19,30.395935058590112,260.0,114.9000015258789,29904.396331787106
-2019-01-15,32192.395141601562,0,123.6999969482422,32192.395141601562
-2019-04-24,91.19587707519531,241.0,133.1999969482422,32192.395141601562
-2019-06-06,30553.59624481201,0,126.4000015258789,30553.59624481201
-2019-06-12,29.196617126464844,244.0,125.0999984741211,30553.59624481201
-2019-06-21,32481.196617126465,0,133.0,32481.196617126465
-2019-07-12,129.19588470458984,240.0,134.8000030517578,32481.196617126465
-2019-07-25,0.39588165283203125,241.0,128.8000030517578,31041.196617126465
-2019-10-28,27715.395881652832,0,115.0,27715.395881652832
-2019-12-19,103.39588165283203,236.0,117.0,27715.395881652832
-2019-12-24,27856.99552154541,0,117.5999984741211,27856.99552154541
-2020-01-31,94.79474639892942,254.0,109.3000030517578,27856.99552154541
-2020-02-14,29431.79474639893,0,115.5,29431.79474639893
-2020-03-18,9.04474639892942,301.0,97.75,29431.79474639893
-2020-04-08,32035.44520568848,0,106.4000015258789,32035.44520568848
-2020-05-20,26.14569854736692,323.0,99.0999984741211,32035.44520568848
-2020-06-12,32423.046684265137,0,100.3000030517578,32423.046684265137
-2020-07-22,44.797210693359375,345.0,93.8499984741211,32423.046684265137
-2020-08-12,34355.04615783692,0,99.4499969482422,34355.04615783692
-2020-09-01,82.14718627929688,337.0,101.6999969482422,34355.04615783692
-2020-10-22,33023.8971862793,0,97.75,33023.8971862793
-2020-11-06,87.5967025756836,317.0,103.9000015258789,33023.8971862793
-2021-01-12,34260.1976699829,0,107.8000030517578,34260.1976699829
-2021-02-08,95.99816131591069,322.0,106.0999984741211,34260.1976699829
-2021-02-19,36771.79865264892,0,113.9000015258789,36771.79865264892
-2021-03-17,2.399131774895068,314.0,117.0999984741211,36771.79865264892
-2021-07-09,35641.399131774895,0,113.5,35641.399131774895
-2021-08-04,74.00009918211435,317.0,112.1999969482422,35641.399131774895
-2021-10-19,32408.000099182114,0,102.0,32408.000099182114
-2021-11-30,81.19905853270393,341.0,94.8000030517578,32408.000099182114
-2022-01-10,32476.199058532704,0,95.0,32476.199058532704
-2022-02-28,24.8985137939344,357.0,90.9000015258789,32476.199058532704
-2022-04-04,33600.74960327147,0,94.0500030517578,33600.74960327147
-2022-04-22,47.74960327146982,356.0,94.25,33600.74960327147
-2022-04-25,33298.15014648436,0,93.4000015258789,33298.15014648436
-2022-06-01,38.65014648436045,342.0,97.25,33298.15014648436
-2022-06-28,32152.450668334946,0,93.9000015258789,32152.450668334946
-2022-08-19,9.201690673809935,335.0,95.9499969482422,32152.450668334946
-2022-08-22,32085.45169067381,0,95.75,32085.45169067381
-2022-08-31,57.350631713852636,347.0,92.3000030517578,32085.45169067381
-2022-09-28,30905.651161193833,0,88.9000015258789,30905.651161193833
-2022-09-29,73.90170288084482,355.0,86.8499984741211,30905.651161193833
-2022-11-25,32804.90061950682,0,92.1999969482422,32804.90061950682
-2022-12-07,14.799591064445849,337.0,97.3000030517578,32804.90061950682
-2022-12-30,36006.40061950682,0,106.8000030517578,36006.40061950682
-2023-03-10,67.20115661619639,352.0,102.0999984741211,36006.40061950682
-2023-03-30,38998.40061950682,0,110.5999984741211,38998.40061950682
-2023-05-25,8.999450683586474,383.0,101.8000030517578,38998.40061950682
+2023-02-27,22.400711059566675,233.0,107.1999969482422,25000.0
