+++ back_test_data/0700_back_test.csv
@@ -1,60 +1,1 @@
 Date,Cash,Number of Stock,Stock Price,Total value
-2015-06-01,23.9990234375,160.0,156.10000610351562,25000.0
-2015-07-24,24263.9990234375,0,151.5,24263.9990234375
-2015-08-06,29.2000732421875,172.0,140.89999389648438,24263.9990234375
-2015-10-14,24556.401123046875,0,142.60000610351562,24556.401123046875
-2015-10-29,136.401123046875,165.0,148.0,24556.401123046875
-2015-11-03,24968.901123046875,0,150.5,24968.901123046875
-2015-12-11,24.500091552734375,169.0,147.60000610351562,24968.901123046875
-2016-01-21,22518.401123046875,0,133.10000610351562,22518.401123046875
-2016-03-21,68.20025634765625,142.0,158.10000610351562,22518.401123046875
-2016-04-07,23043.800689697266,0,161.8000030517578,23043.800689697266
-2016-05-04,23.300689697265625,149.0,154.5,23043.800689697266
-2016-05-25,24533.800689697266,0,164.5,24533.800689697266
-2016-10-31,43.60032653808594,119.0,205.8000030517578,24533.800689697266
-2016-11-24,22915.399963378906,0,192.1999969482422,22915.399963378906
-2016-12-12,148.099212646488,123.0,185.1000061035156,22915.399963378906
-2017-01-09,24206.899963378906,0,195.6000061035156,24206.899963378906
-2017-02-08,134.89996337890625,118.0,204.0,24206.899963378906
-2017-02-16,25150.899963378906,0,212.0,25150.899963378906
-2018-02-12,165.29959106445312,61.0,409.6000061035156,25150.899963378906
-2018-04-16,24748.299591064453,0,403.0,24748.299591064453
-2018-06-26,44.299591064453125,64.0,386.0,24748.299591064453
-2018-11-30,20012.299591064453,0,312.0,20012.299591064453
-2018-12-31,230.29959106445312,63.0,314.0,20012.299591064453
-2019-01-04,19798.099975585938,0,310.6000061035156,19798.099975585938
-2019-01-07,106.89959716796875,62.0,317.6000061035156,19798.099975585938
-2019-01-10,20504.89959716797,0,329.0,20504.89959716797
-2019-05-20,326.89959716796875,59.0,342.0,20504.89959716797
-2019-05-31,19560.89959716797,0,326.0,19560.89959716797
-2019-07-26,310.4989624023401,52.0,370.2000122070313,19560.89959716797
-2019-07-29,19664.89959716797,0,372.2000122070313,19664.89959716797
-2019-08-13,292.89959716796875,58.0,334.0,19664.89959716797
-2019-09-13,20569.699951171875,0,349.6000061035156,20569.699951171875
-2019-10-02,94.699951171875,63.0,325.0,20569.699951171875
-2020-01-13,25697.89956665039,0,406.3999938964844,25697.89956665039
-2020-03-16,62.900482177734375,75.0,341.79998779296875,25697.89956665039
-2020-03-31,28577.90139770508,0,380.2000122070313,28577.90139770508
-2020-04-17,315.50097656250364,69.0,409.6000061035156,28577.90139770508
-2020-04-22,28564.100555419926,0,409.3999938964844,28564.100555419926
-2020-12-28,19.100555419925513,55.0,519.0,28564.100555419926
-2021-01-18,36154.10055541992,0,657.0,36154.10055541992
-2021-02-26,379.1005554199219,54.0,662.5,36154.10055541992
-2021-04-07,34372.10055541992,0,629.5,34372.10055541992
-2021-05-13,152.10055541992188,59.0,580.0,34372.10055541992
-2021-09-06,29947.100555419922,0,505.0,29947.100555419922
-2021-09-29,187.10055541992188,64.0,465.0,29947.100555419922
-2021-12-13,30049.500946044922,0,466.6000061035156,30049.500946044922
-2022-02-17,351.5009460449219,62.0,479.0,30049.500946044922
-2022-04-04,24531.500946044922,0,390.0,24531.500946044922
-2022-04-27,91.10183715820312,73.0,334.79998779296875,24531.500946044922
-2022-06-09,28911.500946044926,0,394.7999877929688,28911.500946044926
-2022-06-30,205.10144042969114,81.0,354.3999938964844,28911.500946044926
-2022-08-26,26465.302429199222,0,324.20001220703125,26465.302429199222
-2022-09-15,65.30242919922239,88.0,300.0,26465.302429199222
-2022-11-11,22804.501892089847,0,258.3999938964844,22804.501892089847
-2022-12-08,269.10232543945676,71.0,317.3999938964844,22804.501892089847
-2022-12-09,23386.702758789066,0,325.6000061035156,23386.702758789066
-2023-02-28,21.902343750003638,68.0,343.6000061035156,23386.702758789066
-2023-03-30,26201.902343750004,0,385.0,26201.902343750004
-2023-04-21,11.901428222659888,75.0,349.20001220703125,26201.902343750004
