+++ constant.py
@@ -1,1 +1,3 @@
 DEFAULT_STOCK_LIST = ['0016','0002','0700','0005']
+BACK_TEST_START_TIME = '2022-11-01'
+BACK_TEST_END_TIME = '2023-10-31'
